+++ src/lib/djerba/genetic_alteration.py
@@ -1,3 +1,14 @@
+"""
+Source of input data for Djerba; corresponds to the genetic_alteration_type metadata field in cBioPortal.
+
+A `genetic_alteration` object reads input from files, database queries, etc.; parses sample-level and gene-level attributes; and returns them in standard data structures.
+
+Includes:
+
+- `genetic alteration`: Abstract base class
+- Subclasses to implement genetic_alteration methods for particular data types
+- `genetic_alteration_factory`: Class to construct an instance of an appropriate `genetic_alteration` subclass, given its name and other configuration data.
+"""
 
 import logging
 import os
@@ -5,18 +16,15 @@ import pandas as pd
 import random
 import tempfile
 import yaml
-from djerba.components import dual_output_component
 from djerba.metrics import mutation_extended_metrics
 from djerba.utilities import constants
 from djerba.utilities.base import base
 from djerba.utilities.tools import system_tools
 
-class genetic_alteration(dual_output_component):
+class genetic_alteration(base):
     """Base class; unit of genetic alteration data for cBioPortal and other reports"""
 
-    # TODO in some instances, could write cBioPortal data to a tempdir and then extract elba metrics
-    # The 'elba' methods get_attributes_for_sample and get_metrics_by_gene should be consistent
-    # and avoid duplication with cBioPortal write_metrics
+    # TODO make into a Python AbstractBaseClass: https://docs.python.org/3/library/abc.html
 
     # top-level config keys
     WORKFLOW_RUN_ID_KEY = 'workflow_run_id'
@@ -33,15 +41,14 @@ class genetic_alteration(dual_output_component):
         try:
             self.genetic_alteration_type = config[constants.GENETIC_ALTERATION_TYPE_KEY]
             self.datatype = config[constants.DATATYPE_KEY]
-            #self.workflow = config[self.WORKFLOW_KEY] # TODO is this field necessary?
             self.metadata = config[self.METADATA_KEY]
             self.input_files = config[self.INPUT_FILES_KEY]
             self.input_directory = config[self.INPUT_DIRECTORY_KEY]
-            self.workflow_run_id = config[self.WORKFLOW_RUN_ID_KEY]
         except KeyError as err:
             self.logger.error("Missing required config key: {0}".format(err))
             raise
-        self.sample_ids = sorted(self.input_files.keys())
+        self.workflow_run_id = config.get(self.WORKFLOW_RUN_ID_KEY, None) # optional param
+        self.sample_ids = self._get_sample_ids()
         self.sample_attributes = self._find_all_sample_attributes()
         # identifier for the genetic_alteration; should be unique in any given config
         self.alteration_id = "%s:%s" % (self.genetic_alteration_type, self.datatype)
@@ -55,6 +62,13 @@ class genetic_alteration(dual_output_component):
         for sample_id in self.sample_ids:
             attributes[sample_id] = {}
         return attributes
+
+    def _get_sample_ids(self):
+        """
+        Find the list of sample IDs. Assumes the input_files config is non-empty.
+        Optionally, can override this method in child classes.
+        """
+        return sorted(self.input_files.keys())
 
     def get_alteration_id(self):
         """ID defined as 'alteration_type:datatype', eg. 'MUTATION_EXTENDED:MAF'"""
@@ -71,7 +85,10 @@ class genetic_alteration(dual_output_component):
         return self.genetic_alteration_type
 
     def get_gene_names(self):
-        """PLACEHOLDER. Get a list of gene names from the input files."""
+        """
+        PLACEHOLDER. Get a list of gene names from the input files.
+        Not run from __init__() because it can be quite slow (eg. reading multiple MAF files).
+        """
         msg = "get_genes method of parent class; not intended for production"
         self.logger.error(msg)
         return []
@@ -93,7 +110,7 @@ class genetic_alteration_factory(base):
     """Supply an instance of the appropriate genetic_alteration subclass for an ALTERATIONTYPE"""
 
     CLASSNAMES = {
-        constants.DEMONSTRATION_TYPE: 'genetic_alteration_demo',
+        constants.CUSTOM_ANNOTATION_TYPE: 'custom_annotation',
         constants.MUTATION_TYPE: 'mutation_extended'
     }
 
@@ -114,43 +131,131 @@ class genetic_alteration_factory(base):
         klass = globals().get(classname)
         return klass(config, study_id, self.log_level, self.log_path)
 
+class custom_annotation(genetic_alteration):
+    """
+    User-defined custom annotation supplied in TSV format.
 
-class genetic_alteration_demo(genetic_alteration):
-    """Dummy class for demonstration and initial testing; not for production use"""
+    ## Input requirements
+
+    - Input is in separate TSV files for gene and sample annotation.
+    - Gene and sample filenames are specified in metadata, as 'gene_tsv' and 'sample_tsv' respectively.
+    - Input files are tab-delimited, and may include comment lines starting with #.
+    - Annotations which contain tab characters may be enclosed in double quotes (").
+    - Column headers in the gene and sample files must be specified in config metadata, as 'gene_headers' and 'sample_headers' respectively.
+    - The first column must be the gene or sample identifier, with header 'Gene' or 'SAMPLE_ID' respectively.
+    - If column headers from metadata are not found in the TSV file, it will raise an error.
+    - Columns in the TSV file which do not appear in metadata are silently ignored.
+
+    ## Notes
+
+    - Useful as a fallback for fields which cannot be automatically obtained by Djerba.
+    - Currently supports Elba output only, not cBioPortal.
+    """
+
+    GENE_HEADERS_KEY = 'gene_headers'
+    GENE_TSV_KEY = 'gene_tsv'
+    SAMPLE_HEADERS_KEY = 'sample_headers'
+    SAMPLE_TSV_KEY = 'sample_tsv'
 
     def _find_all_sample_attributes(self):
-        """DEMONSTRATION METHOD. Return empty attributes"""
-        return {}
-
-    def get_attributes_for_sample(self, sample_id):
-        """DEMONSTRATION METHOD. Get sample-level metrics for Elba"""
-        msg = "get_attributes_for_sample demo method; not intended for production"
-        self.logger.warning(msg)
-        metric_key = ":".join([self.genetic_alteration_type, self.datatype, 'dummy_sample_metric'])
-        attributes = {
-            metric_key: random.randrange(100)
-        }
+        """Read attributes for each sample from a TSV file with specified headers"""
+        try:
+            tsv_path = os.path.join(self.input_directory, self.metadata[self.SAMPLE_TSV_KEY])
+            column_headers = self.metadata[self.SAMPLE_HEADERS_KEY] # must start with SAMPLE_ID_KEY
+        except KeyError as err:
+            self.logger.error("Missing required metadata key: {0}".format(err))
+            raise
+        if column_headers[0] != constants.SAMPLE_ID_KEY:
+            msg = "First sample column header in metadata must be %s" % constants.SAMPLE_ID_KEY
+            self.logger.error(msg)
+            raise ValueError(msg)
+        attributes = {}
+        df = self._read_columns(tsv_path, column_headers, constants.SAMPLE_ID_KEY)
+        for (sample_id, row) in df.iterrows():
+            # row.to_list() is a workaround for int64 conversion; see comments in get_metrics_by_gene
+            values = row.to_list()
+            attributes[sample_id] = {column_headers[i+1]: values[i] for i in range(len(values))}
         return attributes
 
+    def _get_sample_ids(self):
+        """Read sample IDs from TSV input file; overrides method of parent class"""
+        try:
+            tsv_path = os.path.join(self.input_directory, self.metadata[self.SAMPLE_TSV_KEY])
+        except KeyError as err:
+            self.logger.error("Missing required metadata key: {0}".format(err))
+            raise
+        df = self._read_columns(tsv_path, [constants.SAMPLE_ID_KEY], None, index_col=False)
+        return df[constants.SAMPLE_ID_KEY].tolist()
+
+    def _read_columns(self, input_path, column_headers, index_name, index_col=0):
+        """
+        Read specified columns from TSV into a Pandas DataFrame.
+        Raise an error if any requested columns are missing; warn if any values are null.
+        index_col is the parameter of the same name for Pandas read_csv().
+        index_name is the expected name of the index column; must be None if index_col=False.
+        """
+        try:
+            df = pd.read_csv(
+                input_path,
+                delimiter="\t",
+                comment="#",
+                index_col=index_col,
+                usecols=column_headers
+            )
+        except ValueError as err:
+            msg = 'Failed to read TSV from "{0}". Missing required column headers '.format(input_path) +\
+                  'from Djerba config? Pandas error message: "{0}"'.format(err)
+            self.logger.error(msg)
+            raise
+        if df.isnull().values.any():
+            self.logger.warning(
+                'Null values in TSV data read from "{0}" '.format(input_path)+\
+                'with column headers {0}'.format(str(column_headers))
+            )
+        if index_col is False and index_name!=None: # 0 has truth-value False, so use 'is' instead of '=='
+            msg = "index_name must be None if index_col is False"
+            self.logger.error(msg)
+            raise ValueError(msg)
+        elif df.index.name != index_name:
+            msg = "First column header in %s: Expected %s, found %s" % \
+                  (input_path, index_name, df.index.name)
+            self.logger.error(msg)
+            raise ValueError(msg)
+        return df
+
     def get_gene_names(self):
-        """DEMONSTRATION METHOD. Get a list of gene names."""
-        return ["Gene001", "Gene002"]
+        """Find gene names from TSV input file"""
+        if self.gene_names:
+            return self.gene_names
+        try:
+            tsv_path = os.path.join(self.input_directory, self.metadata[self.GENE_TSV_KEY])
+        except KeyError as err:
+            self.logger.error("Missing required metadata key: {0}".format(err))
+            raise
+        df = self._read_columns(tsv_path, [constants.GENE_KEY], None, index_col=False)
+        return df[constants.GENE_KEY].tolist()
 
     def get_metrics_by_gene(self, sample_id):
-        """"DEMONSTRATION METHOD. Get small mutation and indel data for Elba."""
-        msg = "get_metrics_by_gene demo method; not intended for production"
-        self.logger.warning(msg)
-        input_file = self.input_files[sample_id]
-        # generate dummy results as a demonstration
-        metric_key = ":".join([self.alteration_id, 'dummy_metric'])
-        metrics = {}
-        for gene in self.get_gene_names():
-            metrics[gene] = {
-                constants.GENE_KEY: gene,
-                metric_key: random.randrange(101, 1000)
-            }
-        return metrics
-
+        """Read gene-level metrics from input TSV."""
+        try:
+            tsv_path = os.path.join(self.input_directory, self.metadata[self.GENE_TSV_KEY])
+            column_headers = self.metadata[self.GENE_HEADERS_KEY] # must start with GENE_KEY
+        except KeyError as err:
+            self.logger.error("Missing required metadata key: {0}".format(err))
+            raise
+        if column_headers[0] != constants.GENE_KEY:
+            msg = "First gene column header in metadata must be %s" % constants.GENE_KEY
+            self.logger.error(msg)
+            raise ValueError(msg)
+        metrics_by_gene = {}
+        df = self._read_columns(tsv_path, column_headers, constants.GENE_KEY)
+        for (gene_name, row) in df.iterrows():
+            # Can't use row.get() because this may return an int64, which can't be serialized to JSON
+            # row.to_list() converts all values to Python scalars, so can be used as a workaround
+            # See https://bugs.python.org/issue24313
+            values = row.to_list()
+            metrics_by_gene[gene_name] = {column_headers[i+1]: values[i] for i in range(len(values))}
+        return metrics_by_gene
 
 class mutation_extended(genetic_alteration):
     """
@@ -227,13 +332,13 @@ class mutation_extended(genetic_alteration):
         return metrics_by_gene
 
     def write_data(self, out_dir):
-        """cBioPortal. Write mutation data table."""
+        """cBioPortal. Write mutation data table.
 
-        # Read mutation data in MAF format and output in cBioPortal's required MAF format
-        # May enable VCF input at a later date
-        # see https://docs.cbioportal.org/5.1-data-loading/data-loading/file-formats#mutation-data
-        # required modules: vcf2maf/1.6.17, vep/92.0, vep-hg19-cache/92, hg19/p13
-
+        - Read mutation data in MAF format and output in cBioPortal's required MAF format
+        - May enable VCF input at a later date
+        - see https://docs.cbioportal.org/5.1-data-loading/data-loading/file-formats#mutation-data
+        - required modules: vcf2maf/1.6.17, vep/92.0, vep-hg19-cache/92, hg19/p13
+        """
         tmp = tempfile.TemporaryDirectory(prefix='djerba_mutex_')
         tmp_dir = tmp.name
         #tmp_dir = '/scratch2/users/ibancarz/djerba_test/latest' # temporary location for testing
@@ -267,7 +372,7 @@ class mutation_extended(genetic_alteration):
         self.logger.info("Writing concatenated MAF output to %s" % out_path)
         output_df.to_csv(out_path, sep="\t")
         tmp.cleanup()
-        
+
     def write_meta(self, out_dir):
         """cBioPortal. Write mutation metadata."""
         try:
@@ -291,3 +396,7 @@ class mutation_extended(genetic_alteration):
             raise
         with open(os.path.join(out_dir, self.META_FILENAME), 'w') as out_file:
             out_file.write(yaml.dump(meta, sort_keys=True))
+
+    def write(self, out_dir):
+        self.write_data(out_dir)
+        self.write_meta(out_dir)